requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
httplib2==0.9.2
idna==2.6
isort==4.3.16
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.3
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
future==0.17.1
httplib2==0.9.2
idna==2.6
isort==4.3.16
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-debian==0.1.32
python-telegram-bot==11.1.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
EasyProcess==0.2.5
future==0.17.1
heimdall==0.0.6
httplib2==0.9.2
idna==2.6
isort==4.3.16
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.4.1
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-debian==0.1.32
python-telegram-bot==11.1.0
pytz==2018.3
PyVirtualDisplay==0.2.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
